@@ -1,6 +1,7 @@
 """Logic for creating models."""
 from __future__ import annotations as _annotations
 
+import operator
 import sys
 import types
 import typing
@@ -121,17 +122,24 @@ class BaseModel(metaclass=_model_construction.ModelMetaclass):
         __pydantic_serializer__: ClassVar[SchemaSerializer]
         __pydantic_validator__: ClassVar[SchemaValidator]
 
+        model_computed_fields: ClassVar[dict[str, ComputedFieldInfo]]
+        """A dictionary of computed field names and their corresponding `ComputedFieldInfo` objects."""
+
         # Instance attributes
         # Note: we use the non-existent kwarg `init=False` in pydantic.fields.Field below so that @dataclass_transform
         # doesn't think these are valid as keyword arguments to the class initializer.
         __pydantic_extra__: dict[str, Any] | None = _Field(init=False)  # type: ignore
         __pydantic_fields_set__: set[str] = _Field(init=False)  # type: ignore
         __pydantic_private__: dict[str, Any] | None = _Field(init=False)  # type: ignore
+
     else:
         # `model_fields` and `__pydantic_decorators__` must be set for
         # pydantic._internal._generate_schema.GenerateSchema.model_schema to work for a plain BaseModel annotation
         model_fields = {}
+        model_computed_fields = {}
+
         __pydantic_decorators__ = _decorators.DecoratorInfos()
+        __pydantic_parent_namespace__ = None
         # Prevent `BaseModel` from being instantiated directly:
         __pydantic_validator__ = _mock_val_ser.MockValSer(
             'Pydantic models should inherit from BaseModel, BaseModel cannot be instantiated directly',
@@ -150,30 +158,20 @@ class BaseModel(metaclass=_model_construction.ModelMetaclass):
     __pydantic_complete__ = False
     __pydantic_root_model__ = False
 
-    def __init__(__pydantic_self__, **data: Any) -> None:  # type: ignore
+    def __init__(self, /, **data: Any) -> None:  # type: ignore
         """Create a new model by parsing and validating input data from keyword arguments.
 
         Raises [`ValidationError`][pydantic_core.ValidationError] if the input data cannot be
         validated to form a valid model.
 
-        `__init__` uses `__pydantic_self__` instead of the more common `self` for the first arg to
-        allow `self` as a field name.
+        `self` is explicitly positional-only to allow `self` as a field name.
         """
         # `__tracebackhide__` tells pytest and some other tools to omit this function from tracebacks
         __tracebackhide__ = True
-        __pydantic_self__.__pydantic_validator__.validate_python(data, self_instance=__pydantic_self__)
+        self.__pydantic_validator__.validate_python(data, self_instance=self)
 
     # The following line sets a flag that we use to determine when `__init__` gets overridden by the user
-    __init__.__pydantic_base_init__ = True
-
-    @property
-    def model_computed_fields(self) -> dict[str, ComputedFieldInfo]:
-        """Get the computed fields of this model instance.
-
-        Returns:
-            A dictionary of computed field names and their corresponding `ComputedFieldInfo` objects.
-        """
-        return {k: v.info for k, v in self.__pydantic_decorators__.computed_fields.items()}
+    __init__.__pydantic_base_init__ = True  # pyright: ignore[reportFunctionMemberAccess]
 
     @property
     def model_extra(self) -> dict[str, Any] | None:
@@ -211,17 +209,19 @@ class BaseModel(metaclass=_model_construction.ModelMetaclass):
         """
         m = cls.__new__(cls)
         fields_values: dict[str, Any] = {}
-        defaults: dict[str, Any] = {}  # keeping this separate from `fields_values` helps us compute `_fields_set`
+        fields_set = set()
+
         for name, field in cls.model_fields.items():
             if field.alias and field.alias in values:
                 fields_values[name] = values.pop(field.alias)
+                fields_set.add(name)
             elif name in values:
                 fields_values[name] = values.pop(name)
+                fields_set.add(name)
             elif not field.is_required():
-                defaults[name] = field.get_default(call_default_factory=True)
+                fields_values[name] = field.get_default(call_default_factory=True)
         if _fields_set is None:
-            _fields_set = set(fields_values.keys())
-        fields_values.update(defaults)
+            _fields_set = fields_set
 
         _extra: dict[str, Any] | None = None
         if cls.model_config.get('extra') == 'allow':
@@ -237,6 +237,12 @@ class BaseModel(metaclass=_model_construction.ModelMetaclass):
 
         if cls.__pydantic_post_init__:
             m.model_post_init(None)
+            # update private attributes with values set
+            if hasattr(m, '__pydantic_private__') and m.__pydantic_private__ is not None:
+                for k, v in values.items():
+                    if k in m.__private_attributes__:
+                        m.__pydantic_private__[k] = v
+
         elif not cls.__pydantic_root_model__:
             # Note: if there are any private attributes, cls.__pydantic_post_init__ would exist
             # Since it doesn't, that means that `__pydantic_private__` should be set to None
@@ -291,15 +297,15 @@ class BaseModel(metaclass=_model_construction.ModelMetaclass):
 
         Args:
             mode: The mode in which `to_python` should run.
-                If mode is 'json', the dictionary will only contain JSON serializable types.
-                If mode is 'python', the dictionary may contain any Python objects.
+                If mode is 'json', the output will only contain JSON serializable types.
+                If mode is 'python', the output may contain non-JSON-serializable Python objects.
             include: A list of fields to include in the output.
             exclude: A list of fields to exclude from the output.
             by_alias: Whether to use the field's alias in the dictionary key if defined.
             exclude_unset: Whether to exclude fields that have not been explicitly set.
-            exclude_defaults: Whether to exclude fields that are set to their default value from the output.
-            exclude_none: Whether to exclude fields that have a value of `None` from the output.
-            round_trip: Whether to enable serialization and deserialization round-trip support.
+            exclude_defaults: Whether to exclude fields that are set to their default value.
+            exclude_none: Whether to exclude fields that have a value of `None`.
+            round_trip: If True, dumped values should be valid as input for non-idempotent types such as Json[T].
             warnings: Whether to log warnings when invalid fields are encountered.
 
         Returns:
@@ -337,14 +343,14 @@ class BaseModel(metaclass=_model_construction.ModelMetaclass):
 
         Args:
             indent: Indentation to use in the JSON output. If None is passed, the output will be compact.
-            include: Field(s) to include in the JSON output. Can take either a string or set of strings.
-            exclude: Field(s) to exclude from the JSON output. Can take either a string or set of strings.
+            include: Field(s) to include in the JSON output.
+            exclude: Field(s) to exclude from the JSON output.
             by_alias: Whether to serialize using field aliases.
             exclude_unset: Whether to exclude fields that have not been explicitly set.
-            exclude_defaults: Whether to exclude fields that have the default value.
+            exclude_defaults: Whether to exclude fields that are set to their default value.
             exclude_none: Whether to exclude fields that have a value of `None`.
-            round_trip: Whether to use serialization/deserialization between JSON and class instance.
-            warnings: Whether to show any warnings that occurred during serialization.
+            round_trip: If True, dumped values should be valid as input for non-idempotent types such as Json[T].
+            warnings: Whether to log warnings when invalid fields are encountered.
 
         Returns:
             A JSON string representation of the model.
@@ -488,7 +494,7 @@ class BaseModel(metaclass=_model_construction.ModelMetaclass):
 
         Args:
             obj: The object to validate.
-            strict: Whether to raise an exception on invalid fields.
+            strict: Whether to enforce types strictly.
             from_attributes: Whether to extract data from object attributes.
             context: Additional context to pass to the validator.
 
@@ -868,12 +874,50 @@ class BaseModel(metaclass=_model_construction.ModelMetaclass):
             self_type = self.__pydantic_generic_metadata__['origin'] or self.__class__
             other_type = other.__pydantic_generic_metadata__['origin'] or other.__class__
 
-            return (
+            # Perform common checks first
+            if not (
                 self_type == other_type
-                and self.__dict__ == other.__dict__
                 and self.__pydantic_private__ == other.__pydantic_private__
                 and self.__pydantic_extra__ == other.__pydantic_extra__
-            )
+            ):
+                return False
+
+            # We only want to compare pydantic fields but ignoring fields is costly.
+            # We'll perform a fast check first, and fallback only when needed
+            # See GH-7444 and GH-7825 for rationale and a performance benchmark
+
+            # First, do the fast (and sometimes faulty) __dict__ comparison
+            if self.__dict__ == other.__dict__:
+                # If the check above passes, then pydantic fields are equal, we can return early
+                return True
+
+            # We don't want to trigger unnecessary costly filtering of __dict__ on all unequal objects, so we return
+            # early if there are no keys to ignore (we would just return False later on anyway)
+            model_fields = type(self).model_fields.keys()
+            if self.__dict__.keys() <= model_fields and other.__dict__.keys() <= model_fields:
+                return False
+
+            # If we reach here, there are non-pydantic-fields keys, mapped to unequal values, that we need to ignore
+            # Resort to costly filtering of the __dict__ objects
+            # We use operator.itemgetter because it is much faster than dict comprehensions
+            # NOTE: Contrary to standard python class and instances, when the Model class has a default value for an
+            # attribute and the model instance doesn't have a corresponding attribute, accessing the missing attribute
+            # raises an error in BaseModel.__getattr__ instead of returning the class attribute
+            # So we can use operator.itemgetter() instead of operator.attrgetter()
+            getter = operator.itemgetter(*model_fields) if model_fields else lambda _: _utils._SENTINEL
+            try:
+                return getter(self.__dict__) == getter(other.__dict__)
+            except KeyError:
+                # In rare cases (such as when using the deprecated BaseModel.copy() method),
+                # the __dict__ may not contain all model fields, which is how we can get here.
+                # getter(self.__dict__) is much faster than any 'safe' method that accounts
+                # for missing keys, and wrapping it in a `try` doesn't slow things down much
+                # in the common case.
+                self_fields_proxy = _utils.SafeGetItemProxy(self.__dict__)
+                other_fields_proxy = _utils.SafeGetItemProxy(other.__dict__)
+                return getter(self_fields_proxy) == getter(other_fields_proxy)
+
+        # other instance is not a BaseModel
         else:
             return NotImplemented  # delegate to the other item in the comparison
 
@@ -946,26 +990,27 @@ class BaseModel(metaclass=_model_construction.ModelMetaclass):
     # ##### Deprecated methods from v1 #####
     @property
     @typing_extensions.deprecated(
-        'The `__fields__` attribute is deprecated, use `model_fields` instead.', category=PydanticDeprecatedSince20
+        'The `__fields__` attribute is deprecated, use `model_fields` instead.', category=None
     )
     def __fields__(self) -> dict[str, FieldInfo]:
-        warnings.warn('The `__fields__` attribute is deprecated, use `model_fields` instead.', DeprecationWarning)
+        warnings.warn(
+            'The `__fields__` attribute is deprecated, use `model_fields` instead.', category=PydanticDeprecatedSince20
+        )
         return self.model_fields
 
     @property
     @typing_extensions.deprecated(
         'The `__fields_set__` attribute is deprecated, use `model_fields_set` instead.',
-        category=PydanticDeprecatedSince20,
+        category=None,
     )
     def __fields_set__(self) -> set[str]:
         warnings.warn(
-            'The `__fields_set__` attribute is deprecated, use `model_fields_set` instead.', DeprecationWarning
+            'The `__fields_set__` attribute is deprecated, use `model_fields_set` instead.',
+            category=PydanticDeprecatedSince20,
         )
         return self.__pydantic_fields_set__
 
-    @typing_extensions.deprecated(
-        'The `dict` method is deprecated; use `model_dump` instead.', category=PydanticDeprecatedSince20
-    )
+    @typing_extensions.deprecated('The `dict` method is deprecated; use `model_dump` instead.', category=None)
     def dict(  # noqa: D102
         self,
         *,
@@ -976,7 +1021,7 @@ class BaseModel(metaclass=_model_construction.ModelMetaclass):
         exclude_defaults: bool = False,
         exclude_none: bool = False,
     ) -> typing.Dict[str, Any]:  # noqa UP006
-        warnings.warn('The `dict` method is deprecated; use `model_dump` instead.', DeprecationWarning)
+        warnings.warn('The `dict` method is deprecated; use `model_dump` instead.', category=PydanticDeprecatedSince20)
         return self.model_dump(
             include=include,
             exclude=exclude,
@@ -986,9 +1031,7 @@ class BaseModel(metaclass=_model_construction.ModelMetaclass):
             exclude_none=exclude_none,
         )
 
-    @typing_extensions.deprecated(
-        'The `json` method is deprecated; use `model_dump_json` instead.', category=PydanticDeprecatedSince20
-    )
+    @typing_extensions.deprecated('The `json` method is deprecated; use `model_dump_json` instead.', category=None)
     def json(  # noqa: D102
         self,
         *,
@@ -1002,7 +1045,9 @@ class BaseModel(metaclass=_model_construction.ModelMetaclass):
         models_as_dict: bool = PydanticUndefined,  # type: ignore[assignment]
         **dumps_kwargs: Any,
     ) -> str:
-        warnings.warn('The `json` method is deprecated; use `model_dump_json` instead.', DeprecationWarning)
+        warnings.warn(
+            'The `json` method is deprecated; use `model_dump_json` instead.', category=PydanticDeprecatedSince20
+        )
         if encoder is not PydanticUndefined:
             raise TypeError('The `encoder` argument is no longer supported; use field serializers instead.')
         if models_as_dict is not PydanticUndefined:
@@ -1019,18 +1064,18 @@ class BaseModel(metaclass=_model_construction.ModelMetaclass):
         )
 
     @classmethod
-    @typing_extensions.deprecated(
-        'The `parse_obj` method is deprecated; use `model_validate` instead.', category=PydanticDeprecatedSince20
-    )
+    @typing_extensions.deprecated('The `parse_obj` method is deprecated; use `model_validate` instead.', category=None)
     def parse_obj(cls: type[Model], obj: Any) -> Model:  # noqa: D102
-        warnings.warn('The `parse_obj` method is deprecated; use `model_validate` instead.', DeprecationWarning)
+        warnings.warn(
+            'The `parse_obj` method is deprecated; use `model_validate` instead.', category=PydanticDeprecatedSince20
+        )
         return cls.model_validate(obj)
 
     @classmethod
     @typing_extensions.deprecated(
         'The `parse_raw` method is deprecated; if your data is JSON use `model_validate_json`, '
         'otherwise load the data then use `model_validate` instead.',
-        category=PydanticDeprecatedSince20,
+        category=None,
     )
     def parse_raw(  # noqa: D102
         cls: type[Model],
@@ -1044,7 +1089,7 @@ class BaseModel(metaclass=_model_construction.ModelMetaclass):
         warnings.warn(
             'The `parse_raw` method is deprecated; if your data is JSON use `model_validate_json`, '
             'otherwise load the data then use `model_validate` instead.',
-            DeprecationWarning,
+            category=PydanticDeprecatedSince20,
         )
         from .deprecated import parse
 
@@ -1083,7 +1128,7 @@ class BaseModel(metaclass=_model_construction.ModelMetaclass):
     @typing_extensions.deprecated(
         'The `parse_file` method is deprecated; load the data from file, then if your data is JSON '
         'use `model_validate_json`, otherwise `model_validate` instead.',
-        category=PydanticDeprecatedSince20,
+        category=None,
     )
     def parse_file(  # noqa: D102
         cls: type[Model],
@@ -1096,8 +1141,8 @@ class BaseModel(metaclass=_model_construction.ModelMetaclass):
     ) -> Model:
         warnings.warn(
             'The `parse_file` method is deprecated; load the data from file, then if your data is JSON '
-            'use `model_validate_json` otherwise `model_validate` instead.',
-            DeprecationWarning,
+            'use `model_validate_json`, otherwise `model_validate` instead.',
+            category=PydanticDeprecatedSince20,
         )
         from .deprecated import parse
 
@@ -1114,13 +1159,13 @@ class BaseModel(metaclass=_model_construction.ModelMetaclass):
     @typing_extensions.deprecated(
         'The `from_orm` method is deprecated; set '
         "`model_config['from_attributes']=True` and use `model_validate` instead.",
-        category=PydanticDeprecatedSince20,
+        category=None,
     )
     def from_orm(cls: type[Model], obj: Any) -> Model:  # noqa: D102
         warnings.warn(
-            'The `from_orm` method is deprecated; set `model_config["from_attributes"]=True` '
-            'and use `model_validate` instead.',
-            DeprecationWarning,
+            'The `from_orm` method is deprecated; set '
+            "`model_config['from_attributes']=True` and use `model_validate` instead.",
+            category=PydanticDeprecatedSince20,
         )
         if not cls.model_config.get('from_attributes', None):
             raise PydanticUserError(
@@ -1129,15 +1174,17 @@ class BaseModel(metaclass=_model_construction.ModelMetaclass):
         return cls.model_validate(obj)
 
     @classmethod
-    @typing_extensions.deprecated(
-        'The `construct` method is deprecated; use `model_construct` instead.', category=PydanticDeprecatedSince20
-    )
+    @typing_extensions.deprecated('The `construct` method is deprecated; use `model_construct` instead.', category=None)
     def construct(cls: type[Model], _fields_set: set[str] | None = None, **values: Any) -> Model:  # noqa: D102
-        warnings.warn('The `construct` method is deprecated; use `model_construct` instead.', DeprecationWarning)
+        warnings.warn(
+            'The `construct` method is deprecated; use `model_construct` instead.', category=PydanticDeprecatedSince20
+        )
         return cls.model_construct(_fields_set=_fields_set, **values)
 
     @typing_extensions.deprecated(
-        'The copy method is deprecated; use `model_copy` instead.', category=PydanticDeprecatedSince20
+        'The `copy` method is deprecated; use `model_copy` instead. '
+        'See the docstring of `BaseModel.copy` for details about how to handle `include` and `exclude`.',
+        category=None,
     )
     def copy(
         self: Model,
@@ -1161,13 +1208,10 @@ class BaseModel(metaclass=_model_construction.ModelMetaclass):
         ```
 
         Args:
-            include: Optional set or mapping
-                specifying which fields to include in the copied model.
-            exclude: Optional set or mapping
-                specifying which fields to exclude in the copied model.
-            update: Optional dictionary of field-value pairs to override field values
-                in the copied model.
-            deep: If True, the values of fields that are Pydantic models will be deep copied.
+            include: Optional set or mapping specifying which fields to include in the copied model.
+            exclude: Optional set or mapping specifying which fields to exclude in the copied model.
+            update: Optional dictionary of field-value pairs to override field values in the copied model.
+            deep: If True, the values of fields that are Pydantic models will be deep-copied.
 
         Returns:
             A copy of the model with included, excluded and updated fields as specified.
@@ -1175,7 +1219,7 @@ class BaseModel(metaclass=_model_construction.ModelMetaclass):
         warnings.warn(
             'The `copy` method is deprecated; use `model_copy` instead. '
             'See the docstring of `BaseModel.copy` for details about how to handle `include` and `exclude`.',
-            DeprecationWarning,
+            category=PydanticDeprecatedSince20,
         )
         from .deprecated import copy_internals
 
@@ -1214,29 +1258,29 @@ class BaseModel(metaclass=_model_construction.ModelMetaclass):
         return copy_internals._copy_and_set_values(self, values, fields_set, extra, private, deep=deep)
 
     @classmethod
-    @typing_extensions.deprecated(
-        'The `schema` method is deprecated; use `model_json_schema` instead.', category=PydanticDeprecatedSince20
-    )
+    @typing_extensions.deprecated('The `schema` method is deprecated; use `model_json_schema` instead.', category=None)
     def schema(  # noqa: D102
         cls, by_alias: bool = True, ref_template: str = DEFAULT_REF_TEMPLATE
     ) -> typing.Dict[str, Any]:  # noqa UP006
-        warnings.warn('The `schema` method is deprecated; use `model_json_schema` instead.', DeprecationWarning)
+        warnings.warn(
+            'The `schema` method is deprecated; use `model_json_schema` instead.', category=PydanticDeprecatedSince20
+        )
         return cls.model_json_schema(by_alias=by_alias, ref_template=ref_template)
 
     @classmethod
     @typing_extensions.deprecated(
         'The `schema_json` method is deprecated; use `model_json_schema` and json.dumps instead.',
-        category=PydanticDeprecatedSince20,
+        category=None,
     )
     def schema_json(  # noqa: D102
         cls, *, by_alias: bool = True, ref_template: str = DEFAULT_REF_TEMPLATE, **dumps_kwargs: Any
     ) -> str:  # pragma: no cover
-        import json
-
         warnings.warn(
             'The `schema_json` method is deprecated; use `model_json_schema` and json.dumps instead.',
-            DeprecationWarning,
+            category=PydanticDeprecatedSince20,
         )
+        import json
+
         from .deprecated.json import pydantic_encoder
 
         return json.dumps(
@@ -1246,44 +1290,47 @@ class BaseModel(metaclass=_model_construction.ModelMetaclass):
         )
 
     @classmethod
-    @typing_extensions.deprecated(
-        'The `validate` method is deprecated; use `model_validate` instead.', category=PydanticDeprecatedSince20
-    )
+    @typing_extensions.deprecated('The `validate` method is deprecated; use `model_validate` instead.', category=None)
     def validate(cls: type[Model], value: Any) -> Model:  # noqa: D102
-        warnings.warn('The `validate` method is deprecated; use `model_validate` instead.', DeprecationWarning)
+        warnings.warn(
+            'The `validate` method is deprecated; use `model_validate` instead.', category=PydanticDeprecatedSince20
+        )
         return cls.model_validate(value)
 
     @classmethod
     @typing_extensions.deprecated(
         'The `update_forward_refs` method is deprecated; use `model_rebuild` instead.',
-        category=PydanticDeprecatedSince20,
+        category=None,
     )
     def update_forward_refs(cls, **localns: Any) -> None:  # noqa: D102
         warnings.warn(
-            'The `update_forward_refs` method is deprecated; use `model_rebuild` instead.', DeprecationWarning
+            'The `update_forward_refs` method is deprecated; use `model_rebuild` instead.',
+            category=PydanticDeprecatedSince20,
         )
         if localns:  # pragma: no cover
             raise TypeError('`localns` arguments are not longer accepted.')
         cls.model_rebuild(force=True)
 
     @typing_extensions.deprecated(
-        'The private method `_iter` will be removed and should no longer be used.', category=PydanticDeprecatedSince20
+        'The private method `_iter` will be removed and should no longer be used.', category=None
     )
     def _iter(self, *args: Any, **kwargs: Any) -> Any:
-        warnings.warn('The private method `_iter` will be removed and should no longer be used.', DeprecationWarning)
-
+        warnings.warn(
+            'The private method `_iter` will be removed and should no longer be used.',
+            category=PydanticDeprecatedSince20,
+        )
         from .deprecated import copy_internals
 
         return copy_internals._iter(self, *args, **kwargs)
 
     @typing_extensions.deprecated(
         'The private method `_copy_and_set_values` will be removed and should no longer be used.',
-        category=PydanticDeprecatedSince20,
+        category=None,
     )
     def _copy_and_set_values(self, *args: Any, **kwargs: Any) -> Any:
         warnings.warn(
-            'The private method  `_copy_and_set_values` will be removed and should no longer be used.',
-            DeprecationWarning,
+            'The private method `_copy_and_set_values` will be removed and should no longer be used.',
+            category=PydanticDeprecatedSince20,
         )
         from .deprecated import copy_internals
 
@@ -1292,26 +1339,26 @@ class BaseModel(metaclass=_model_construction.ModelMetaclass):
     @classmethod
     @typing_extensions.deprecated(
         'The private method `_get_value` will be removed and should no longer be used.',
-        category=PydanticDeprecatedSince20,
+        category=None,
     )
     def _get_value(cls, *args: Any, **kwargs: Any) -> Any:
         warnings.warn(
-            'The private method  `_get_value` will be removed and should no longer be used.', DeprecationWarning
+            'The private method `_get_value` will be removed and should no longer be used.',
+            category=PydanticDeprecatedSince20,
         )
-
         from .deprecated import copy_internals
 
         return copy_internals._get_value(cls, *args, **kwargs)
 
     @typing_extensions.deprecated(
         'The private method `_calculate_keys` will be removed and should no longer be used.',
-        category=PydanticDeprecatedSince20,
+        category=None,
     )
     def _calculate_keys(self, *args: Any, **kwargs: Any) -> Any:
         warnings.warn(
-            'The private method `_calculate_keys` will be removed and should no longer be used.', DeprecationWarning
+            'The private method `_calculate_keys` will be removed and should no longer be used.',
+            category=PydanticDeprecatedSince20,
         )
-
         from .deprecated import copy_internals
 
         return copy_internals._calculate_keys(self, *args, **kwargs)
@@ -1359,18 +1406,20 @@ def create_model(  # noqa: C901
     __slots__: tuple[str, ...] | None = None,
     **field_definitions: Any,
 ) -> type[Model]:
-    """Dynamically creates and returns a new Pydantic model, in other words, `create_model` dynamically creates a
+    """Usage docs: https://docs.pydantic.dev/2.6/concepts/models/#dynamic-model-creation
+
+    Dynamically creates and returns a new Pydantic model, in other words, `create_model` dynamically creates a
     subclass of [`BaseModel`][pydantic.BaseModel].
 
     Args:
         __model_name: The name of the newly created model.
         __config__: The configuration of the new model.
         __doc__: The docstring of the new model.
-        __base__: The base class for the new model.
-        __module__: The name of the module that the model belongs to,
-            if `None` the value is taken from `sys._getframe(1)`
+        __base__: The base class or classes for the new model.
+        __module__: The name of the module that the model belongs to;
+            if `None`, the value is taken from `sys._getframe(1)`
         __validators__: A dictionary of methods that validate fields.
-        __cls_kwargs__: A dictionary of keyword arguments for class creation.
+        __cls_kwargs__: A dictionary of keyword arguments for class creation, such as `metaclass`.
         __slots__: Deprecated. Should not be passed to `create_model`.
         **field_definitions: Attributes of the new model. They should be passed in the format:
             `<name>=(<type>, <default value>)` or `<name>=(<type>, <FieldInfo>)`.
