@@ -682,50 +682,50 @@ table_rows: list[Row] = [
         tuple,
         strict=True,
         python_input=True,
-        core_schemas=[core_schema.TuplePositionalSchema, core_schema.TupleVariableSchema],
+        core_schemas=[core_schema.TupleSchema],
     ),
     Row(
         tuple,
         'Array',
         strict=True,
         json_input=True,
-        core_schemas=[core_schema.TuplePositionalSchema, core_schema.TupleVariableSchema],
+        core_schemas=[core_schema.TupleSchema],
     ),
     Row(
         tuple,
         list,
         python_input=True,
-        core_schemas=[core_schema.TuplePositionalSchema, core_schema.TupleVariableSchema],
+        core_schemas=[core_schema.TupleSchema],
     ),
     Row(
         tuple,
         set,
         python_input=True,
-        core_schemas=[core_schema.TuplePositionalSchema, core_schema.TupleVariableSchema],
+        core_schemas=[core_schema.TupleSchema],
     ),
     Row(
         tuple,
         frozenset,
         python_input=True,
-        core_schemas=[core_schema.TuplePositionalSchema, core_schema.TupleVariableSchema],
+        core_schemas=[core_schema.TupleSchema],
     ),
     Row(
         tuple,
         deque,
         python_input=True,
-        core_schemas=[core_schema.TuplePositionalSchema, core_schema.TupleVariableSchema],
+        core_schemas=[core_schema.TupleSchema],
     ),
     Row(
         tuple,
         'dict_keys',
         python_input=True,
-        core_schemas=[core_schema.TuplePositionalSchema, core_schema.TupleVariableSchema],
+        core_schemas=[core_schema.TupleSchema],
     ),
     Row(
         tuple,
         'dict_values',
         python_input=True,
-        core_schemas=[core_schema.TuplePositionalSchema, core_schema.TupleVariableSchema],
+        core_schemas=[core_schema.TupleSchema],
     ),
     Row(
         set,
