@@ -216,20 +216,21 @@ class WrapValidator:
 if TYPE_CHECKING:
 
     class _OnlyValueValidatorClsMethod(Protocol):
-        def __call__(self, __cls: Any, __value: Any) -> Any:
+        def __call__(self, cls: Any, value: Any, /) -> Any:
             ...
 
     class _V2ValidatorClsMethod(Protocol):
-        def __call__(self, __cls: Any, __input_value: Any, __info: _core_schema.ValidationInfo) -> Any:
+        def __call__(self, cls: Any, value: Any, info: _core_schema.ValidationInfo, /) -> Any:
             ...
 
     class _V2WrapValidatorClsMethod(Protocol):
         def __call__(
             self,
-            __cls: Any,
-            __input_value: Any,
-            __validator: _core_schema.ValidatorFunctionWrapHandler,
-            __info: _core_schema.ValidationInfo,
+            cls: Any,
+            value: Any,
+            handler: _core_schema.ValidatorFunctionWrapHandler,
+            info: _core_schema.ValidationInfo,
+            /,
         ) -> Any:
             ...
 
@@ -379,7 +380,10 @@ class ModelWrapValidatorHandler(_core_schema.ValidatorFunctionWrapHandler, Proto
     """@model_validator decorated function handler argument type. This is used when `mode='wrap'`."""
 
     def __call__(  # noqa: D102
-        self, input_value: Any, outer_location: str | int | None = None
+        self,
+        value: Any,
+        outer_location: str | int | None = None,
+        /,
     ) -> _ModelTypeCo:  # pragma: no cover
         ...
 
@@ -395,8 +399,9 @@ class ModelWrapValidatorWithoutInfo(Protocol[_ModelType]):
         # this can be a dict, a model instance
         # or anything else that gets passed to validate_python
         # thus validators _must_ handle all cases
-        __value: Any,
-        __handler: ModelWrapValidatorHandler[_ModelType],
+        value: Any,
+        handler: ModelWrapValidatorHandler[_ModelType],
+        /,
     ) -> _ModelType:
         ...
 
@@ -410,10 +415,27 @@ class ModelWrapValidator(Protocol[_ModelType]):
         # this can be a dict, a model instance
         # or anything else that gets passed to validate_python
         # thus validators _must_ handle all cases
-        __value: Any,
-        __handler: ModelWrapValidatorHandler[_ModelType],
-        __info: _core_schema.ValidationInfo,
+        value: Any,
+        handler: ModelWrapValidatorHandler[_ModelType],
+        info: _core_schema.ValidationInfo,
+        /,
     ) -> _ModelType:
+        ...
+
+
+class FreeModelBeforeValidatorWithoutInfo(Protocol):
+    """A @model_validator decorated function signature.
+    This is used when `mode='before'` and the function does not have info argument.
+    """
+
+    def __call__(  # noqa: D102
+        self,
+        # this can be a dict, a model instance
+        # or anything else that gets passed to validate_python
+        # thus validators _must_ handle all cases
+        value: Any,
+        /,
+    ) -> Any:
         ...
 
 
@@ -428,7 +450,23 @@ class ModelBeforeValidatorWithoutInfo(Protocol):
         # this can be a dict, a model instance
         # or anything else that gets passed to validate_python
         # thus validators _must_ handle all cases
-        __value: Any,
+        value: Any,
+        /,
+    ) -> Any:
+        ...
+
+
+class FreeModelBeforeValidator(Protocol):
+    """A `@model_validator` decorated function signature. This is used when `mode='before'`."""
+
+    def __call__(  # noqa: D102
+        self,
+        # this can be a dict, a model instance
+        # or anything else that gets passed to validate_python
+        # thus validators _must_ handle all cases
+        value: Any,
+        info: _core_schema.ValidationInfo,
+        /,
     ) -> Any:
         ...
 
@@ -442,8 +480,9 @@ class ModelBeforeValidator(Protocol):
         # this can be a dict, a model instance
         # or anything else that gets passed to validate_python
         # thus validators _must_ handle all cases
-        __value: Any,
-        __info: _core_schema.ValidationInfo,
+        value: Any,
+        info: _core_schema.ValidationInfo,
+        /,
     ) -> Any:
         ...
 
@@ -457,7 +496,9 @@ ModelAfterValidator = Callable[[_ModelType, _core_schema.ValidationInfo], _Model
 """A `@model_validator` decorated function signature. This is used when `mode='after'`."""
 
 _AnyModelWrapValidator = Union[ModelWrapValidator[_ModelType], ModelWrapValidatorWithoutInfo[_ModelType]]
-_AnyModeBeforeValidator = Union[ModelBeforeValidator, ModelBeforeValidatorWithoutInfo]
+_AnyModeBeforeValidator = Union[
+    FreeModelBeforeValidator, ModelBeforeValidator, FreeModelBeforeValidatorWithoutInfo, ModelBeforeValidatorWithoutInfo
+]
 _AnyModelAfterValidator = Union[ModelAfterValidator[_ModelType], ModelAfterValidatorWithoutInfo[_ModelType]]
 
 
@@ -499,7 +540,7 @@ def model_validator(
 
     Example usage:
     ```py
-    from typing import Optional
+    from typing_extensions import Self
 
     from pydantic import BaseModel, ValidationError, model_validator
 
@@ -508,7 +549,7 @@ def model_validator(
         height: float
 
         @model_validator(mode='after')
-        def verify_square(self) -> 'Rectangle':
+        def verify_square(self) -> Self:
             if self.width != self.height:
                 raise ValueError('width and height do not match')
             return self
@@ -523,8 +564,7 @@ def model_validator(
         print(e)
         '''
         1 validation error for Square
-        __root__
-          width and height do not match (type=value_error)
+          Value error, width and height do not match [type=value_error, input_value={'width': 1, 'height': 2}, input_type=dict]
         '''
     ```
 
