@@ -1,6 +1,5 @@
 import platform
 import re
-import sys
 from datetime import date, timedelta
 from pathlib import Path
 from types import SimpleNamespace
@@ -31,17 +30,12 @@ from pydantic.functional_serializers import model_serializer
 from pydantic.json_schema import JsonSchemaValue
 from pydantic.type_adapter import TypeAdapter
 
-if sys.version_info < (3, 11):
-    from typing_extensions import get_overloads
-else:
-    from typing import get_overloads
-
 
 def deprecated_from_orm(model_type: Type[BaseModel], obj: Any) -> Any:
     with pytest.warns(
         PydanticDeprecatedSince20,
         match=re.escape(
-            'The `from_orm` method is deprecated; set `model_config["from_attributes"]=True` '
+            "The `from_orm` method is deprecated; set `model_config['from_attributes']=True` "
             'and use `model_validate` instead.'
         ),
     ):
@@ -330,6 +324,14 @@ def test_fields_set():
     match = '^The `__fields_set__` attribute is deprecated, use `model_fields_set` instead.'
     with pytest.warns(PydanticDeprecatedSince20, match=match):
         assert m.__fields_set__ == {'x'}
+
+
+def test_fields_dir():
+    class Model(BaseModel):
+        x: int
+        y: int = 2
+
+    assert '__fields__' not in dir(Model)
 
 
 @pytest.mark.parametrize('attribute,value', [('allow', 'allow'), ('ignore', 'ignore'), ('forbid', 'forbid')])
@@ -689,7 +691,7 @@ def test_update_forward_refs():
 def test_copy_and_set_values():
     m = SimpleModel(x=1)
     with pytest.warns(
-        PydanticDeprecatedSince20, match='^The private method  `_copy_and_set_values` will be removed and '
+        PydanticDeprecatedSince20, match='^The private method `_copy_and_set_values` will be removed and '
     ):
         m2 = m._copy_and_set_values(values={'x': 2}, fields_set={'x'}, deep=False)
 
@@ -698,7 +700,7 @@ def test_copy_and_set_values():
 
 def test_get_value():
     m = SimpleModel(x=1)
-    with pytest.warns(PydanticDeprecatedSince20, match='^The private method  `_get_value` will be removed and '):
+    with pytest.warns(PydanticDeprecatedSince20, match='^The private method `_get_value` will be removed and '):
         v = m._get_value(
             [1, 2, 3],
             to_dict=False,
@@ -718,49 +720,51 @@ def test_deprecated_module(tmp_path: Path) -> None:
 
     assert hasattr(parse_obj_as, '__deprecated__')
     with pytest.warns(
-        PydanticDeprecatedSince20, match='parse_obj_as is deprecated. Use pydantic.TypeAdapter.validate_python instead.'
+        PydanticDeprecatedSince20,
+        match='`parse_obj_as` is deprecated. Use `pydantic.TypeAdapter.validate_python` instead.',
     ):
         parse_obj_as(Model, {'x': 1})
 
     assert hasattr(schema_json_of, '__deprecated__')
     with pytest.warns(
-        PydanticDeprecatedSince20, match='schema_json_of is deprecated. Use pydantic.TypeAdapter.json_schema instead.'
+        PydanticDeprecatedSince20,
+        match='`schema_json_of` is deprecated. Use `pydantic.TypeAdapter.json_schema` instead.',
     ):
         schema_json_of(Model)
 
     assert hasattr(schema_of, '__deprecated__')
     with pytest.warns(
-        PydanticDeprecatedSince20, match='schema_of is deprecated. Use pydantic.TypeAdapter.json_schema instead.'
+        PydanticDeprecatedSince20, match='`schema_of` is deprecated. Use `pydantic.TypeAdapter.json_schema` instead.'
     ):
         schema_of(Model)
 
     assert hasattr(load_str_bytes, '__deprecated__')
-    with pytest.warns(PydanticDeprecatedSince20, match='load_str_bytes is deprecated.'):
+    with pytest.warns(PydanticDeprecatedSince20, match='`load_str_bytes` is deprecated.'):
         load_str_bytes('{"x": 1}')
 
     assert hasattr(load_file, '__deprecated__')
     file = tmp_path / 'main.py'
     file.write_text('{"x": 1}')
-    with pytest.warns(PydanticDeprecatedSince20, match='load_file is deprecated.'):
+    with pytest.warns(PydanticDeprecatedSince20, match='`load_file` is deprecated.'):
         load_file(file)
 
     assert hasattr(pydantic_encoder, '__deprecated__')
     with pytest.warns(
-        PydanticDeprecatedSince20, match='pydantic_encoder is deprecated, use BaseModel.model_dump instead.'
+        PydanticDeprecatedSince20,
+        match='`pydantic_encoder` is deprecated, use `pydantic_core.to_jsonable_python` instead.',
     ):
         pydantic_encoder(Model(x=1))
 
     assert hasattr(custom_pydantic_encoder, '__deprecated__')
     with pytest.warns(
-        PydanticDeprecatedSince20, match='custom_pydantic_encoder is deprecated, use BaseModel.model_dump instead.'
+        PydanticDeprecatedSince20, match='`custom_pydantic_encoder` is deprecated, use `BaseModel.model_dump` instead.'
     ):
         custom_pydantic_encoder({int: lambda x: str(x)}, Model(x=1))
 
     assert hasattr(timedelta_isoformat, '__deprecated__')
-    with pytest.warns(PydanticDeprecatedSince20, match='timedelta_isoformat is deprecated.'):
+    with pytest.warns(PydanticDeprecatedSince20, match='`timedelta_isoformat` is deprecated.'):
         timedelta_isoformat(timedelta(seconds=1))
 
-    assert all(hasattr(func, '__deprecated__') for func in get_overloads(validate_arguments))
     with pytest.warns(
         PydanticDeprecatedSince20, match='The `validate_arguments` method is deprecated; use `validate_call` instead.'
     ):
