@@ -38,7 +38,6 @@ from pydantic import (
     GenerateSchema,
     GetCoreSchemaHandler,
     PrivateAttr,
-    PydanticDeprecatedSince20,
     PydanticUndefinedAnnotation,
     PydanticUserError,
     SecretStr,
@@ -651,9 +650,11 @@ def test_hash_function_works_when_instance_dict_modified():
     assert hash(m) == h
 
     # Keys can be missing, e.g. when using the deprecated copy method.
-    # This should change the hash, and more importantly hashing shouldn't raise a KeyError.
+    # This could change the hash, and more importantly hashing shouldn't raise a KeyError
+    # We don't assert here, because a hash collision is possible: the hash is not guaranteed to change
+    # However, hashing must not raise an exception, which simply calling hash() checks for
     del m.__dict__['a']
-    assert h != hash(m)
+    hash(m)
 
 
 def test_default_hash_function_overrides_default_hash_function():
@@ -1532,7 +1533,7 @@ def test_model_export_inclusion_inheritance():
         s4: str = 'v4'
 
     class Parent(BaseModel):
-        # b will be included since fields are set idependently
+        # b will be included since fields are set independently
         model_config = ConfigDict(fields={'b': {'include': ...}})
         a: int
         b: int
@@ -2976,12 +2977,9 @@ def test_deferred_core_schema() -> None:
 
 
 def test_help(create_module):
-    # since pydoc/help access all attributes to generate their documentation,
-    # this triggers the deprecation warnings.
-    with pytest.warns(PydanticDeprecatedSince20):
-        module = create_module(
-            # language=Python
-            """
+    module = create_module(
+        # language=Python
+        """
 import pydoc
 
 from pydantic import BaseModel
@@ -2992,8 +2990,7 @@ class Model(BaseModel):
 
 help_result_string = pydoc.render_doc(Model)
 """
-        )
-
+    )
     assert 'class Model' in module.help_result_string
 
 
