@@ -591,12 +591,9 @@ def test_json_schema_extra_on_model_and_on_field():
 
 
 def test_help(create_module):
-    # since pydoc/help access all attributes to generate their documentation,
-    # this triggers the deprecation warnings.
-    with pytest.warns(PydanticDeprecatedSince20):
-        module = create_module(
-            # language=Python
-            """
+    module = create_module(
+        # language=Python
+        """
 import pydoc
 
 from pydantic import RootModel
@@ -604,8 +601,7 @@ from pydantic import RootModel
 
 help_result_string = pydoc.render_doc(RootModel)
 """
-        )
-
+    )
     assert 'class RootModel' in module.help_result_string
 
 
