@@ -1,9 +1,10 @@
 import datetime
-from typing import Dict, List, Tuple, TypeVar, Union
+from dataclasses import dataclass
+from typing import Dict, Generic, List, Tuple, TypeVar, Union
 
 import pytest
 from annotated_types import MaxLen
-from typing_extensions import Annotated, TypeAliasType
+from typing_extensions import Annotated, Literal, TypeAliasType
 
 from pydantic import BaseModel, Field, ValidationError
 from pydantic.type_adapter import TypeAdapter
@@ -131,6 +132,18 @@ def test_recursive_type_alias() -> None:
             }
         },
     }
+
+
+def test_recursive_type_alias_name():
+    T = TypeVar('T')
+
+    @dataclass
+    class MyGeneric(Generic[T]):
+        field: T
+
+    MyRecursiveType = TypeAliasType('MyRecursiveType', Union[MyGeneric['MyRecursiveType'], int])
+    json_schema = TypeAdapter(MyRecursiveType).json_schema()
+    assert sorted(json_schema['$defs'].keys()) == ['MyGeneric_MyRecursiveType_', 'MyRecursiveType']
 
 
 def test_type_alias_annotated() -> None:
@@ -353,3 +366,25 @@ def test_redefined_type_alias():
 
     data = {'inner': {'x': 'hello'}, 'y': 1}
     assert MyOuterModel.model_validate(data).model_dump() == data
+
+
+def test_type_alias_to_type_with_ref():
+    class Div(BaseModel):
+        type: Literal['Div'] = 'Div'
+        components: List['AnyComponent']
+
+    AnyComponent = TypeAliasType('AnyComponent', Div)
+
+    adapter = TypeAdapter(AnyComponent)
+    adapter.validate_python({'type': 'Div', 'components': [{'type': 'Div', 'components': []}]})
+    with pytest.raises(ValidationError) as exc_info:
+        adapter.validate_python({'type': 'Div', 'components': [{'type': 'NotDiv', 'components': []}]})
+    assert exc_info.value.errors(include_url=False) == [
+        {
+            'ctx': {'expected': "'Div'"},
+            'input': 'NotDiv',
+            'loc': ('components', 0, 'type'),
+            'msg': "Input should be 'Div'",
+            'type': 'literal_error',
+        }
+    ]
