@@ -38,9 +38,9 @@ from uuid import UUID
 import annotated_types
 import dirty_equals
 import pytest
-from dirty_equals import HasRepr, IsOneOf, IsStr
+from dirty_equals import HasRepr, IsFloatNan, IsOneOf, IsStr
 from pydantic_core import CoreSchema, PydanticCustomError, SchemaError, core_schema
-from typing_extensions import Annotated, Literal, TypedDict, get_args
+from typing_extensions import Annotated, Literal, NotRequired, TypedDict, get_args
 
 from pydantic import (
     UUID1,
@@ -48,6 +48,7 @@ from pydantic import (
     UUID4,
     UUID5,
     AfterValidator,
+    AllowInfNan,
     AwareDatetime,
     Base64Bytes,
     Base64Str,
@@ -64,6 +65,8 @@ from pydantic import (
     FutureDate,
     FutureDatetime,
     GetCoreSchemaHandler,
+    GetPydanticSchema,
+    ImportString,
     InstanceOf,
     Json,
     JsonValue,
@@ -76,20 +79,24 @@ from pydantic import (
     NonNegativeInt,
     NonPositiveFloat,
     NonPositiveInt,
+    OnErrorOmit,
     PastDate,
     PastDatetime,
     PositiveFloat,
     PositiveInt,
     PydanticInvalidForJsonSchema,
+    PydanticSchemaGenerationError,
     SecretBytes,
     SecretStr,
     SerializeAsAny,
     SkipValidation,
+    Strict,
     StrictBool,
     StrictBytes,
     StrictFloat,
     StrictInt,
     StrictStr,
+    StringConstraints,
     Tag,
     TypeAdapter,
     ValidationError,
@@ -107,9 +114,6 @@ from pydantic import (
     validate_call,
 )
 from pydantic.dataclasses import dataclass as pydantic_dataclass
-from pydantic.errors import PydanticSchemaGenerationError
-from pydantic.functional_validators import AfterValidator
-from pydantic.types import AllowInfNan, GetPydanticSchema, ImportString, Strict, StringConstraints
 
 try:
     import email_validator
@@ -873,6 +877,8 @@ def test_string_import_callable(annotation):
     [
         ('math:cos', 'math.cos', 'json'),
         ('math:cos', math.cos, 'python'),
+        ('math.cos', 'math.cos', 'json'),
+        ('math.cos', math.cos, 'python'),
         pytest.param(
             'os.path', 'posixpath', 'json', marks=pytest.mark.skipif(sys.platform == 'win32', reason='different output')
         ),
@@ -897,6 +903,22 @@ def test_string_import_any(value: Any, expected: Any, mode: Literal['json', 'pyt
         thing: ImportString
 
     assert PyObjectModel(thing=value).model_dump(mode=mode) == {'thing': expected}
+
+
+@pytest.mark.parametrize(
+    ('value', 'validate_default', 'expected'),
+    [
+        (math.cos, True, math.cos),
+        ('math:cos', True, math.cos),
+        (math.cos, False, math.cos),
+        ('math:cos', False, 'math:cos'),
+    ],
+)
+def test_string_import_default_value(value: Any, validate_default: bool, expected: Any):
+    class PyObjectModel(BaseModel):
+        thing: ImportString = Field(default=value, validate_default=validate_default)
+
+    assert PyObjectModel().thing == expected
 
 
 @pytest.mark.parametrize('value', ['oss', 'os.os', f'{__name__}.x'])
@@ -1456,9 +1478,9 @@ def test_datetime_errors(DatetimeModel):
     # insert_assert(exc_info.value.errors(include_url=False))
     assert exc_info.value.errors(include_url=False) == [
         {
-            'type': 'datetime_parsing',
+            'type': 'datetime_from_date_parsing',
             'loc': ('dt',),
-            'msg': 'Input should be a valid datetime, month value is outside expected range of 1-12',
+            'msg': 'Input should be a valid datetime or date, month value is outside expected range of 1-12',
             'input': '2017-13-05T19:47:07',
             'ctx': {'error': 'month value is outside expected range of 1-12'},
         },
@@ -2529,13 +2551,34 @@ def test_float_validation():
     ]
 
 
-def test_finite_float_validation():
+def test_infinite_float_validation():
     class Model(BaseModel):
         a: float = None
 
     assert Model(a=float('inf')).a == float('inf')
     assert Model(a=float('-inf')).a == float('-inf')
     assert math.isnan(Model(a=float('nan')).a)
+
+
+@pytest.mark.parametrize(
+    ('ser_json_inf_nan', 'input', 'output', 'python_roundtrip'),
+    (
+        ('null', float('inf'), 'null', None),
+        ('null', float('-inf'), 'null', None),
+        ('null', float('nan'), 'null', None),
+        ('constants', float('inf'), 'Infinity', float('inf')),
+        ('constants', float('-inf'), '-Infinity', float('-inf')),
+        ('constants', float('nan'), 'NaN', IsFloatNan),
+    ),
+)
+def test_infinite_float_json_serialization(ser_json_inf_nan, input, output, python_roundtrip):
+    class Model(BaseModel):
+        model_config = ConfigDict(ser_json_inf_nan=ser_json_inf_nan)
+        a: float
+
+    json_string = Model(a=input).model_dump_json()
+    assert json_string == f'{{"a":{output}}}'
+    assert json.loads(json_string) == {'a': python_roundtrip}
 
 
 @pytest.mark.parametrize('value', [float('inf'), float('-inf'), float('nan')])
@@ -4400,6 +4443,8 @@ def test_frozenset_field_not_convertible():
         ('1.5 M', int(1.5e6), '1.4MiB', '1.5MB'),
         ('5.1kib', 5222, '5.1KiB', '5.2KB'),
         ('6.2EiB', 7148113328562451456, '6.2EiB', '7.1EB'),
+        ('8bit', 1, '1B', '1B'),
+        ('1kbit', 125, '125B', '125B'),
     ),
 )
 def test_bytesize_conversions(input_value, output, human_bin, human_dec):
@@ -4423,6 +4468,8 @@ def test_bytesize_to():
     assert m.size.to('MiB') == pytest.approx(1024)
     assert m.size.to('MB') == pytest.approx(1073.741824)
     assert m.size.to('TiB') == pytest.approx(0.0009765625)
+    assert m.size.to('bit') == pytest.approx(8589934592)
+    assert m.size.to('kbit') == pytest.approx(8589934.592)
 
 
 def test_bytesize_raises():
@@ -4442,6 +4489,9 @@ def test_bytesize_raises():
     m = Model(size='1MB')
     with pytest.raises(PydanticCustomError, match='byte unit'):
         m.size.to('bad_unit')
+
+    with pytest.raises(PydanticCustomError, match='byte unit'):
+        m.size.to('1ZiB')
 
 
 def test_deque_success():
@@ -6099,3 +6149,68 @@ def test_json_value():
             'type': 'invalid-json-value',
         }
     ]
+
+
+def test_json_value_with_subclassed_types():
+    class IntType(int):
+        pass
+
+    class FloatType(float):
+        pass
+
+    class StrType(str):
+        pass
+
+    class ListType(list):
+        pass
+
+    class DictType(dict):
+        pass
+
+    adapter = TypeAdapter(JsonValue)
+    valid_json_data = {'int': IntType(), 'float': FloatType(), 'str': StrType(), 'list': ListType(), 'dict': DictType()}
+    assert adapter.validate_python(valid_json_data) == valid_json_data
+
+
+def test_json_value_roundtrip() -> None:
+    # see https://github.com/pydantic/pydantic/issues/8175
+    class MyModel(BaseModel):
+        val: Union[str, JsonValue]
+
+    round_trip_value = json.loads(MyModel(val=True).model_dump_json())['val']
+    assert round_trip_value is True, round_trip_value
+
+
+def test_on_error_omit() -> None:
+    OmittableInt = OnErrorOmit[int]
+
+    class MyTypedDict(TypedDict):
+        a: NotRequired[OmittableInt]
+        b: NotRequired[OmittableInt]
+
+    class Model(BaseModel):
+        a_list: List[OmittableInt]
+        a_dict: Dict[OmittableInt, OmittableInt]
+        a_typed_dict: MyTypedDict
+
+    actual = Model(
+        a_list=[1, 2, 'a', 3],
+        a_dict={1: 1, 2: 2, 'a': 'a', 'b': 0, 3: 'c', 4: 4},
+        a_typed_dict=MyTypedDict(a=1, b='xyz'),  # type: ignore
+    )
+
+    expected = Model(a_list=[1, 2, 3], a_dict={1: 1, 2: 2, 4: 4}, a_typed_dict=MyTypedDict(a=1))
+
+    assert actual == expected
+
+
+def test_on_error_omit_top_level() -> None:
+    ta = TypeAdapter(OnErrorOmit[int])
+
+    assert ta.validate_python(1) == 1
+    assert ta.validate_python('1') == 1
+
+    # we might want to just raise the OmitError or convert it to a ValidationError
+    # if it hits the top level, but this documents the current behavior at least
+    with pytest.raises(SchemaError, match='Uncaught Omit error'):
+        ta.validate_python('a')
